Added config file and template dir of example files
