+++ bin/categories/TEMPLATE/example_program/install.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+# TODO: Installation script for PROGRAM goes here
