+++ bin/categories/TEMPLATE/run_sat.json
@@ -1,0 +1,17 @@
+{
+  "TIMEOUT" : "timeout here in seconds, e.g. 30.0",
+
+  "PROGRAMS" :
+  {
+    "example_program":
+    {
+      "command_nickname1": "path to tool with command line parameters, including timeout",
+      "command_nickname2": "path to tool with command line parameters, including timeout"
+    },
+    "example_program2":
+    {
+      "command_nickname1": "path to tool with command line parameters, including timeout",
+      "command_nickname2": "path to tool with command line parameters, including timeout"
+    }
+  }
+}
