+++ bin/categories/sat/run_sat.json
@@ -1,5 +1,4 @@
 {
-  "FILE_EXTENSION" : "cnf",
   "TIMEOUT" : 30.0,
 
   "PROGRAMS" :
