--- bin/categories/smt/cvc4/install.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-wget -O tools/smt/cvc4 "http://cvc4.cs.stanford.edu/downloads/builds/x86_64-linux-opt/cvc4-1.6-x86_64-linux-opt"
-chmod 755 tools/smt/cvc4
