--- bin/categories/smt/run_smt.json
@@ -1,16 +1,0 @@
-{
-  "FILE_EXTENSION" : "smt2",
-  "TIMEOUT" : 30.0,
-
-  "PROGRAMS" :
-  {
-    "cvc4":
-    {
-      "cvc4_stock": "tools/smt/cvc4 --lang smt --strings-exp --tlimit=33000 -q"
-    },
-    "z3":
-    {
-      "z3_stock": "tools/smt/z3 -T:33"
-    }
-  }
-}
